add CountyPoolsIF fix calculation of CountyPoolsWithStyles.totalCards
